pack-blit do not preserve downscaling factor state
